requirements.txt
requests==2.31.0
supabase==1.0.3
python-dotenv==1.0.1 